+++ src/collective/cover/tests/test_cover.py
@@ -7,12 +7,10 @@ from collective.cover.testing import INTEGRATION_TESTING
 from plone import api
 from plone.app.dexterity.behaviors.exclfromnav import IExcludeFromNavigation
 from plone.app.lockingbehavior.behaviors import ILocking
-from plone.app.referenceablebehavior.referenceable import IReferenceable
 from plone.app.testing import setRoles
 from plone.app.testing import TEST_USER_ID
 from plone.dexterity.interfaces import IDexterityFTI
 from plone.registry.interfaces import IRegistry
-from plone.uuid.interfaces import IAttributeUUID
 from zope.component import createObject
 from zope.component import getUtility
 from zope.component import queryUtility
@@ -57,10 +55,6 @@ class CoverIntegrationTestCase(unittest.TestCase):
 
     def test_locking_behavior(self):
         self.assertTrue(ILocking.providedBy(self.cover))
-
-    def test_is_referenceable(self):
-        self.assertTrue(IReferenceable.providedBy(self.cover))
-        self.assertTrue(IAttributeUUID.providedBy(self.cover))
 
     def test_cover_selectable_as_folder_default_view(self):
         self.folder.setDefaultPage('c1')